+++ demo/demo.js
@@ -14,7 +14,7 @@ function init () {
   scene.getCamera().setTarget(0, 0, 0);
   scene.getCamera().setPosition(0, 0, 4);
   // scene.setClearColor([1, 1, 1, 0]);
-  scene.setLightPosition([4.0, 4.0, 4.0]);
+  scene.setLightPosition([40.0, 40.0, 40.0]);
   // scene.setAmbientColor([0.4, 0.4, 0.4]);
 
   var gui = createGUI(scene);
@@ -108,7 +108,7 @@ function init () {
   // box.material.wireframe = true;
   // box.material.wireframeWidth = 2;
   // box.material.wireframeOnly = false;
-  box.material.light = false;
+  // box.material.light = false;
   // box.material.doubleSided = true;
   // box.setScale(1.0, 1.0, 1.0);
   // box.setPosition(0.0, 1.7, 0.0);
@@ -120,7 +120,7 @@ function init () {
     wrapS: 'REPEAT',
     wrapT: 'REPEAT'
   };
-  box.material.diffuseColor = [0, 0, 1, 1];
+  // box.material.diffuseColor = [0, 0, 1, 1];
   // box.material.transparent = true;
   // box.material.transparency = 0.5;
   // box.outline = true;