+++ src/VertexArray.js
@@ -3,10 +3,17 @@
  * @param {WebGLRenderingContext} gl WebGLRenderingContext
  * @param {Object} options
  * @example
- *     buffers: { position: [], normal: [], uv: [], color: [], index: [] },
+ *     buffers: {
+ *       position: [],
+ *       normal: [],
+ *       uv: [],
+ *       tangent: [],
+ *       color: [],
+ *       index: [],
+ *       targets: [{ position: [], normal: [], tangent: [] }]
+ *     },
  *     offset: 0,
- *     mode: 'TRIANGLES',
- *     instancedAttrs: null // ['offset']
+ *     mode: 'TRIANGLES'
  */
 var VertexArray = wg.VertexArray = function (gl, options) {
   var self = this,
@@ -15,23 +22,14 @@ var VertexArray = wg.VertexArray = function (gl, options) {
   self._gl = gl;
   self._vao = gl.createVertexArray();
   self._bufferMap = {};
+  self._program = null;
 
   gl.bindVertexArray(self._vao);
   Object.keys(buffers).forEach(function (attrName) {
-    var attribute = attributesMap[attrName];
-    if (!attribute && attrName !== 'index') {
-      return;
-    }
     var bufferData = buffers[attrName];
     var bufferObject = gl.createBuffer();
     self._bufferMap[attrName] = bufferObject;
     var element_type, element_size, array;
-
-    if (attrName === 'position') {
-      if (!self._index) {
-        self._count = bufferData.length / attribute.size;
-      }
-    }
 
     if (attrName === 'color') {
       self._color = true;
@@ -59,10 +57,32 @@ var VertexArray = wg.VertexArray = function (gl, options) {
       self._element_size = element_size;
       self._count = bufferData.length;
       gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, array, gl.STATIC_DRAW);
+    } else if (attrName === 'targets') {
+      self._bufferMap.targets = {};
+      // TODO targets.length > 4
+      buffers.targets.forEach(function (target, index) {
+        createBuffer('position' + index, target.position);
+        createBuffer('normal' + index, target.normal);
+        createBuffer('tangent' + index, target.tangent);
+
+        function createBuffer (name, data) {
+          if (data) {
+            var targetBufferObject = gl.createBuffer();
+            self._bufferMap[name] = targetBufferObject;
+            gl.bindBuffer(gl.ARRAY_BUFFER, targetBufferObject);
+            gl.bufferData(gl.ARRAY_BUFFER, data, gl.STATIC_DRAW);
+          }
+        }
+      });
     } else {
       gl.bindBuffer(gl.ARRAY_BUFFER, bufferObject);
       gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(bufferData), gl.STATIC_DRAW);
-      self._setBufferOptions(attribute, bufferData, options.instancedAttrs && options.instancedAttrs.indexOf(attrName) >= 0);
+      if (attrName === 'position' && !self._index) {
+        self._count = bufferData.length / 3;
+      }
+      if (attrName === 'offset') {
+        self._instanceCount = bufferData.length / 16;
+      }
     }
   });
   gl.bindVertexArray(null);
@@ -76,61 +96,77 @@ VertexArray.prototype.setPosition = function (datas) {
   this.setBufferDatas('position', datas);
 };
 
-VertexArray.prototype.setBufferDatas = function (name, datas, instanced) {
+VertexArray.prototype.setBufferDatas = function (name, datas) {
   var self = this,
     gl = self._gl,
-    bufferObject, attribute;
+    bufferObject;
   gl.bindVertexArray(self._vao);
   bufferObject = self._bufferMap[name];
   if (!bufferObject) {
-    attribute = attributesMap[name];
-    if (!attribute) {
-      console.error('Unknown attribute: ' + name);
-    }
     self._bufferMap[name] = bufferObject = gl.createBuffer();
-    gl.bindBuffer(gl.ARRAY_BUFFER, bufferObject);
-    gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(datas), gl.STATIC_DRAW);
-    self._setBufferOptions(attribute, datas, instanced);
-  } else {
-    gl.bindBuffer(gl.ARRAY_BUFFER, bufferObject);
-    gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(datas), gl.STATIC_DRAW);
+  }
+  gl.bindBuffer(gl.ARRAY_BUFFER, bufferObject);
+  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(datas), gl.STATIC_DRAW);
+  if (name === 'offset') {
+    self._instanceCount = datas.length / 16;
+  }
+  if (name === 'position' && !self._index) {
+    self._count = datas.length / 3;
   }
 };
 
-VertexArray.prototype._setBufferOptions = function (attribute, datas, instanced) {
+VertexArray.prototype._setBufferOptions = function (attribute, buffer) {
   //https://stackoverflow.com/questions/38853096/webgl-how-to-bind-values-to-a-mat4-attribute
   var self = this,
     gl = self._gl,
     attributeCount, i;
+  gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
   if (attribute.size > 4) {
     attributeCount = attribute.size / 4;
     for (i = 0; i < attributeCount; i++) {
-      gl.enableVertexAttribArray(attribute.index + i);
-      gl.vertexAttribPointer(attribute.index + i, 4, gl.FLOAT, false, 4 * attribute.size, 16 * i);
-      if (instanced) {
-        gl.vertexAttribDivisor(attribute.index + i, 1);
+      gl.enableVertexAttribArray(attribute.location + i);
+      gl.vertexAttribPointer(attribute.location + i, 4, gl.FLOAT, false, 4 * attribute.size, 16 * i);
+      if (attribute.name === 'offset') {
+        gl.vertexAttribDivisor(attribute.location + i, 1);
       }
     }
   } else {
-    gl.enableVertexAttribArray(attribute.index);
+    gl.enableVertexAttribArray(attribute.location);
     // index, size, type, normalized, stride, offset
-    gl.vertexAttribPointer(attribute.index, attribute.size, gl.FLOAT, false, 0, 0);
-    if (instanced) {
-      gl.vertexAttribDivisor(attribute.index, 1);
+    gl.vertexAttribPointer(attribute.location, attribute.size, gl.FLOAT, false, 0, 0);
+    if (attribute.name === 'offset') {
+      gl.vertexAttribDivisor(attribute.location, 1);
     }
   }
-  if (instanced) {
-    self._instanceCount = datas.length / attribute.size;
+};
+
+VertexArray.prototype.bind = function (program) {
+  var self = this,
+    gl = self._gl,
+    bufferMap = self._bufferMap;
+  gl.bindVertexArray(self._vao);
+  if (self._program !== program) {
+    self._program = program;
+    Object.keys(bufferMap).forEach(function (key) {
+      var attribute = program._attributes[key],
+        buffer = bufferMap[key];
+      if (attribute && key !== 'index') {
+        self._setBufferOptions(attribute, bufferMap[key]);
+      }
+    });
   }
 };
 
 VertexArray.prototype.draw = function (preDrawCallback) {
   var self = this,
     gl = self._gl;
-  gl.bindVertexArray(self._vao);
   if (self._parts) {
     self._parts.forEach(function (part) {
+      if (part.material.transparent !== gl._transparent) {
+        return;
+      }
       preDrawCallback && preDrawCallback(part);
+      self.bind(gl._program);
       // TODO instance
       part.counts.forEach(function (item) {
         if (self._index) {
@@ -141,6 +177,7 @@ VertexArray.prototype.draw = function (preDrawCallback) {
       });
     });
   } else {
+    self.bind(gl._program);
     // TODO check if draw*Instanced is supported
     if (self._instanceCount) {
       if (self._index) {
@@ -162,11 +199,13 @@ VertexArray.prototype.draw = function (preDrawCallback) {
 };
 
 VertexArray.prototype.dispose = function () {
-  var self = this;
-  Object.keys(self._bufferMap).forEach(function (key) {
-    self._gl.deleteBuffer(self._bufferMap[key]);
+  var self = this,
+    gl = self._gl,
+    bufferMap = self._bufferMap;
+  Object.keys(bufferMap).forEach(function (key) {
+    gl.deleteBuffer(bufferMap[key]);
   });
-  self._gl.deleteVertexArray(self._vao);
+  gl.deleteVertexArray(self._vao);
   self._vao = null;
   self._bufferMap = {};
 };